Moved more utility code
